[flow] Clean up unused to_merge_or_check
Summary:
Now that to_merge contains all dependent files, it is equivalent to
to_merge_or_check. This diff consolidates all users to a single set and deletes
the other one.

Note that all of this cleanup is preparation for a bigger change, where I will
consolidate most of these sets into a single abstraction, for both clarity and
perf.

Reviewed By: panagosg7

Differential Revision: D50871990

fbshipit-source-id: d4e5f1ea700954f210900c6dadf83eeb2645e8fa