[flow] make TypeDestructorTriggerT a use_t
Summary:
With the previous diff we can make this now a use_t.

Unfortunately it's hard to completely match existing checker behavior, since a lot of it depends on the ordering of upper bounds. Moving `TypeDestructorTriggerT` from `t` to `use_t` perturbs this ordering. One place where this is pronounced is when considering upper bounds for the solution to implicit instantiation. To match existing semantics and impose a smaller error diff, when considering upper bounds in `t_of_use_t`, I'm arranging upper bounds so that we consider types first, then `TypeDestructorTriggerT`, and then the rest of the `use_t`s.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D50765610

fbshipit-source-id: 3a57b64147a817ace9d99ad29c9aefff383cf78e