[flow] move ErrorCollation profiling to types_js.ml
Summary:
This diff moves incremental error collation profiling to types_js.ml, so that it accounts for some filtering that happens there. This filtering is necessary for the whole process and did not happen in the pre-incremental version.

There's still some more operations that are not accounted for:
* local errors
* final unioning of the local & merge errors

I believe both of these operations are relatively cheap, and can more safely be excluded from this profiling.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D50913763

fbshipit-source-id: f520e15648209ecf368385a179a7785419c3e844