[perf] Rewrite environment management in `name_resolver` to avoid quadratic behavior
Summary:
## Background

In this diff, I significantly rewrite the way we handle environment in `name_resolver`, with the goal of eliminating the quadratic behavior. Before we go into implementation detail, let's first review the status quo:

We have two things that we both call `env` right now:

1. `env` in `name_resolver` state. This is a mapping from names to state of names (undeclared, undefined, write, refinement, etc). This environment has a mutable interior (`val_ref` and `heap_refinement` for every `env_val`). We also push and pop this environment when we enter a new scope.
2. `Env.t` we generate ad-hoc. This is related to 1, but it's completely immutable. It represents a snapshot of the environment at some point. Usually, we generate and save this environment for the purpose of later merging it back into the main environment in 1.

When we do refinement invalidation or create an pristine environment when we enter the function body, we will "havoc". In this process, we will do something to every single name in the `env_state.env`. When we have a lot of functions , this will cause quadratic behavior. For example, Flow takes 4s to run just name_resolver on [this file](https://github.com/facebook/metro/blob/fcd8867d5b2ccf4133e1c12f224695e315d09b6f/flow-typed/npm/babel-types_v7.x.x.js#L3194), while actual checking of this file is relatively fast.

## Intuition of a fix

Let's examine closely what's happening when we enter the function body. For every name, we will reset the value to havoc value if it's not const like, or it has not been initialized or we have not encounter the declaration.

I claim that this is only one way to look at what needs to be done. The high level goal is not to havoc some value when entering the function body, but to create an environment that everything is assumed to be initialized. If something is already const initialized and refined, we are happy to take that as well. e.g.

```
declare const canBeRefined: string | null;
if (canBeRefined == null) throw ''
const f = () => {
  canBeRefined; // will read refined value
  refBeforeDecl; // will read the initialized write
}
const refBeforeDecl = 3;
```

The intuition is that, we don't know what state the names will be by the time we are in the function body, so we just assume that everything is in it's initial more general state. Inside and outside of the function body, from the perspective of `name_resolver`, **is two completely unconnected world that can't possibly affect each other**.

The above sentence is the key to the optimization idea. We will treat the external world as frozen when we enter the function body, so when we are invalidating refinements, we are absolutely certain that those external values won't be affected, and do not have to be touched at all. Of course, we might read the external environment, but we can keep tracked of these captured values separately, and only pay the cost of havoc for what's newly declared in the function, and what we actually capture in the function body.

## Implementation

### `module FullEnv`

In this diff, I introduced a new module called `FullEnv` that models the global name -> value mapping. This replaces the `env_val Nel.t SMap.t` in `env_state.env` previously. This module defines a opaque type to enforce the idea by construction that, once you enter the function body, all the environment stacked below it is completely frozen. The module provides no ways to change the environment below, but only provide functions to change the current function scope.

Internally, `FullEnv` is a stack of function scope. Inside the function scope, we have `env_value Nel.t SMap.t` as usual. For context, the `Nel.t` part represents the shadowing stack of the same name. Previously, this is global, but now it will be local to the function scope. Here is an example that illustrates the differences:

```
const foo = 1; // val 1
function () {
  const foo = 1; // val 2
  {
    /*
     * Previously, env_state.env = {"foo" => [val3; val2; val1], ...other globals}
     * now, env_state.env =
     * [{"foo" => [val3; val2]}] :: [{"foo" => [val1}] :: [(other globals)]
     */
    const foo = 1; // val 3
  }
}
```

However, this is not the full picture, since we do need to capture values from environment below. Therefore, we separately keep track of a map of captured names to their values *local to the function*. Unlike the old setup, this map will be populated lazily as we traverse the program.

Hence, internally, `FullEnv.t` is modeled as `function_scope Nel.t`, where `function_scope = { local_stacked_env: env_val Nel.t SMap.t; captured: env_val SMap.t ref }`.

### Lazily populating captured values

When we read a name, we will of course first check whether it exists in `local_stacked_env`. If so, we behave the same as before. If not, then we must be capturing a value, so we need to copy it over from the environment stack below.

The "copy" is not a literal byte-by-byte copy. Instead, we need to respect the behavior that we read a "pristine" value that's initialized to havoc or a read of refined const. What exactly to copy is informed by the `should_havoc_val_to_initialized` lambda.

Since the populating `captured` map is lazy, we no longer have the guarantee that different environment and their snapshots always have the same key set, which makes merging environment hard. The complexity will be explained in the section below.

### `module PartialEnvSnapshot` and environment merging

This is previously the `Env` module. It got the "snapshot" suffix to distinguish between a global mutable environment vs an immutable snapshot. It's now partial because we no longer snapshot the entire environment, but only the function scope top in the `FullEnv` stack.

One result of being partial is that, during environment merging, we no longer have the invariant that the current function scope has the same set of keys as a partial env snapshot. Instead, we have the invariant that the current environment's keys is a super set of environment snapshot, since snapshot captures the environment earlier in time, but the current mutable function scope's keys are always monotonously increasing as we visit the program and lazily populate captured values. Hence, when we are merging environments, if we encounter a key that's not in the env snapshot, we need to read it from below.

## Conclusion

How does it help perf? We only pay for what we actually capture in the function scope, and as a result, create much smaller snapshot. This diff is not supposed to introduce behavior differences, only perf improvements.

Changelog: [perf] We rewrite the way we analyze syntactic environment required for type checking. On a large file like [this](https://github.com/facebook/metro/blob/fcd8867d5b2ccf4133e1c12f224695e315d09b6f/flow-typed/npm/babel-types_v7.x.x.js#L3194), it improves the checking time from 4s to 2s.

Reviewed By: jbrown215

Differential Revision: D50850606

fbshipit-source-id: 08c0e936652114ce3a6a5b64221f307c8c4ee7fd