Add telemetry/ping LSP request
Summary:
LSP clients can send this request to the server and the server will respond as
quickly as possible with an empty response.

The request is handled in parallel by the server, but it's latency is affected
by server availability. This is by design: we can use this request to measure
server availability, as is done later in this stack.

Reviewed By: panagosg7

Differential Revision: D50911728

fbshipit-source-id: 84a8067b7fd88be28b4b204c8e5d7a0756f17654