[flow][cleanup][EZ] Trivialize trust representation
Summary:
We are probably never going to ship trusted type in the current form. In this stack, I start to delete things related to trusted types.

This diff trivialize the trust representation to be just unit, since all the call sites that care about its value are gone.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D51090678

fbshipit-source-id: 97c39b1acc7d58b7dfc64dbaa788e8ac1f332896