[flow] Make useRef return an opaque type
Summary:
We're planning to begin enforcing React rules around refs within components, and in order to do so, we need to know whether something is the result of a call to `useRef` (as opposed to an arbitrary object with a current property).

To do so, this diff makes the result of calling `useRef` an opaque type. In future diffs we'll use that fact to perform enforcement.

As a result of adding new opaque types, we also need to improve our support of refining opaque types to avoid breaking some existing code.

Reviewed By: SamChou19815

Differential Revision: D49745457

fbshipit-source-id: 9c46551db4a133ebd2a6f9d5a143be9ad70ed284