[flow][multiplatform] Typing behavior for `@supportsPlatform` pragma
Summary:
This is the diff that ties everything together, which adds complete type checking using `supportsPlatform`. It can be seen as a generalization of the current multi-platform support.

`supportsPlatform` gives people the capability to declare a set of available platforms for every single module, instead of inferring it from platform extensions or blanket assumption that every single module needs to support all possible platforms.

Previously, when we do multi-platform validation, we do the following:

1. For every common interface file, we require that platform specific impl file exist for every platform.
2. For every platform specific file, we require that it conforms to the common interface file.

Now, we need some generalization:

1. We no longer require that platform specific file exist for every platform, if things like `supportsPlatform` says that it only needs to work for a subset of platforms
2. If the available platform set of the implementation file and common interface file do not intersect, then the platform specific file doesn't need to conform to that common interface.
3. We do additional validation that file can only import another file if the available platform set of the current file is a subset of the imported module's available platform set.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D50773132

fbshipit-source-id: fc9d509792429fafb8e9a022ed53a3109105ad86