[flow][multiplatform][util] Add platform set util that wraps the raw bitset API to prepare for multiplatform work
Summary:
We want to compactly represent a set of platforms supported by a module in the multiplatform world. We can use a bitset so that we don't have to store a string set.

The previous diff adds the bitset util. This diff adds another util that wraps the raw bitset API, so that the mutable nature of bitset and its low level representation is not leaked, and we can expose only necessary high level API to the caller.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D50773131

fbshipit-source-id: c8558adc6b0f785d0601a8406cc4bab97c2d09a1