[flow][multiplatform] Parsing support for `@supportsPlatform` pragma
Summary:
In this diff, I implemented the parsing support for `supportsPlatform` pragma, which lets user declare an arbitrary subset of platforms a logical module intends to support. This is a repeatable pragma, which means that if you want to express the idea of a module supporting both ios and android, you would write:

```
/**
 * supportsPlatform ios
 * supportsPlatform android
 */
```

This pragma is completely ignored if multiplatform flag is not on. If the multiplatform flag is on, we will ensure that you don't declare random platforms that's not in the flowconfig.

Since this is the start of the diff stack, I would like to discuss the philosophical question first. Why are we doing this instead of inferring it from the existence of platform files. Eventually we can do both, but I think it would be necessary that there exists a way to declare these things explicitly. Therefore, I decide to start with the explicit solution, and depending on user feedback and actual usage, decide whether to follow up with a more implicit system.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D50773138

fbshipit-source-id: 9b7513b3fb89f808e6e101512af7b290f4adfa9c