[flow][perf] Use OpenT indirection for `Flow_js_utils.Instantiation_kit.mk_distributive_tparam_subst_fn`
Summary:
Similar to D43012554, but we do it for substitution when we are dealing with a distributive tparam in conditional type and mapped type. This helps to avoid deep traversal of large/deep objects.

Changelog: [perf] We optimized substitution performance when conditional type and mapped types are operating on large and deep objects.

Reviewed By: panagosg7

Differential Revision:
D51138416

------------------------------------------------------------------------
(from 5eda4007851de8a5015462e84d832cea56f64a60)

fbshipit-source-id: 78fa978c2247a1bbfe1a3e18d3853874a7691fa3