[flow] include "init libs" in server status
Summary:
The server status (shown in the bottom status bar of vscode) is not always updated when a new stage of checking has initiated. This has as a result to often show the latest message that was sent to the IDE, even though a new stage of typechecking may have initiated.

This diff fills in this gap in the server status for the library initialization stage. Before, this message was missing and it would appear as if parsing was stuck in the last batch, while Flow was processing libdefs.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D51149772

fbshipit-source-id: 06a1827941be5f6f49a6c9f4fbccc16b6ad2a4ff