[flow] Enforce that refs are not read from directly within component bodies
Summary:
This diff builds the initial version of the rule against reading refs from within component bodies as per the rules of React: https://react.dev/reference/react/useRef#avoiding-recreating-the-ref-contents. For now, this just handles direct results of calls to `useRef`, and it only detects reads within the toplevel of the component: it doesn't for example, error on cases like
```
component Component(items: any) {
  const ref = useRef(null);
  const renderItem = (item: any) => {
    const current = ref.current;
    return <Foo item={item} current={current} />;
  };
  return <Items>{items.map((item) => renderItem(item))}</Items>;
}
```
These errors will be handled in a future diff that changes the approach of this analysis. For now, we just perform a traversal over the typed ast, finding components and then inspecting the bodies of the components (skipping any nested functions). We determine whether a value is a result of useRef by seeing if its type is an opaque type that matches the ID of useRefResult

Reviewed By: jbrown215

Differential Revision: D50279980

fbshipit-source-id: 035fe8b5e1414d00eac419136431e9b69255f831