[flow][ide] Support find references and rename for private names
Summary:
The references are computed along with definition location in a single pass in `GetDefUtils`. Even in global find ref mode, we will skip the normal map-reduce flow of recheck -> find refs, but instead directly return the results.

Renaming is a little tricky. What happens if you rename a private property to non-private? This is important, since the initial rename text field in vscode will be populated with # prefixed names. I decide to follow the typescript behavior: if the new name starts with `#`, it will continue to be private name, if not, it will be changed to normal identifier.

Changelog: [ide] Find references and rename now work for private names

Reviewed By: panagosg7

Differential Revision: D51274750

fbshipit-source-id: 823dcbe1dec37963fbe904adb903e7c4544c90ea