[flow][jsdoc] Collect documentation from declare export statements
Summary:
I thought that hover docs don't work for libdefs, but it turns out that it has nothing to do with libdef, but because the docs searcher doesn't work for `declare export` statements.

e.g. In

```
/** docs */
declare export function foo(): void;
```

The comment belongs to the `declare export` statement, but the doc search logic is only implemented for `declare_function` statement. Similar to how we treat export statements, we can fix it by forwarding comments to the wrapped statement.

Changelog: [ide][fix] Documentation will now show up for `declare export` statements in hover and autocomplete. This means that jsdoc for exports in `declare module` will finally show up in IDE services.

Reviewed By: panagosg7

Differential Revision: D51316417

fbshipit-source-id: b7fc728afddc53067b5f50b6d2132f93fb0dee49