[flow] fix speculation unsoundness with spread arg
Summary:
In code like following
```
declare function foo<TArgs>(...args: TArgs): TArgs;
declare var args: Array<string> | Array<string>;

const x = foo(...args);
x as Array<number>;
```
Flow creates a type variable E to represent the element of the array that represents the spread arguments to the call. Even though Flow considers the reason for the type of the type parameter `TArgs` as an instantiation variable, the reason for the array element type E was not. Instead, we were using a custom reason.

This decision has an effect on how we later handle the union check. Before this change we would record the tvar E as [OpenUnresolved](https://github.com/facebook/flow/blob/38c13f46ae28377e637fe4054ef0b10eecbeb0ae/src/typing/resolvableTypeJob.ml#L89). When type variables like this are not type parameters that we are solving for, they are [resolved to the any type](https://github.com/facebook/flow/blob/38c13f46ae28377e637fe4054ef0b10eecbeb0ae/src/typing/speculation_kit.ml#L357).

This causes the cast in the last line to trivially succeed.

With this change `W` is considered as related to the instantiation, and so it is not pinned to `any`.

Reviewed By: SamChou19815

Differential Revision: D50962106

fbshipit-source-id: b147951b5bd9df3f1d1984730063854b375c452c