[flow] Generate full typed AST for libdef files
Summary:
Before we can provide IDE services for libdefs, we must first generate a full typed AST that matches the shape of the original untyped AST.

To maintain the current behavior of completely ignoring statements that shouldn't be in libdefs (e.g. while), I will generate two aloc ASTs. The filter one will be passed to env_builder so that we don't generate write entries for things in ignored statements, and the other AST that will be passed to statement.ml. To ensure that we actually ignore those statements that should be ignored, I put an error mapped typed AST nodes for those statements in node_cache.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D51328128

fbshipit-source-id: 08a00be25f2ec92897e95f6f91398331f17278d3