[flow] add Ty_normalizer_debug module with no dependency on Flow_js
Summary:
This diff creates a version of the normalizer intended for debug purposes. It can be called from within the Flow_js module and it's recursive deps.

To do this we fill in the parts that Ty_normalizer.Make expects with functions that do not depend on Flow_js. Specifically:
* We do not evaluate eval types or $Keys, and return the "non-evaluated" version instead.
* We do not evaluate typeapps, and return the un-elaborated type application form instead
* We compute builtins directly without dependency on Flow_js.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D51365464

fbshipit-source-id: afbf4c67ba410438198ec41485c30855067c0759