[flow] mitigate skipped characters on Linked Editing Range requests
Summary:
[Linked Editing Range](https://microsoft.github.io/language-server-protocol/specifications/lsp/3.17/specification/#textDocument_linkedEditingRange) is an LSP feature that allows the server given an input range to determine all ranges that should have the same content. In Flow we use in code like
```
<Foo> </Foo>
```
to update the opening and closing tag at the same time, as the user edits the opening or closing tag.

It was recently observed that it was possible, by typing the characters of the opening tag quickly to throw off the synchronization on the closing end.

The reason for this is that by the time the Flow server responds to the latest linkedEditingRange request, the response it returns is invalid.

Fixing this behavior so that it is impossible to have this race condition is probably not going to be easy, especially given that the feature is built on string representations of ranges. Note, that this bug exists even in TS, though it does take more to trigger it compared to Flow (mainly a larger file).

Instead, this diff attempts to mitigate the problem by allowing the server to respond to linkedEditingRange requests when the tags of the same JSX elements don't match. The rationale is that this will allow the server to recover the tag of the side under edit on the other side, when the range it provides is a valid one.

Before, this was not possible since the server would not respond to this request when the file had a parse error, and mismatched tags is a parse error. With this change, the server will proceed with responding to linkedEditingRange requests when parse errors are just ExpectedJSXClosingTag ones.

This could be extended to other parse errors as well, but we might end up updating tags other than the ones expected by the user.

Changelog: [fix] Mitigates a bug in support for Linked Editing Range where typing characters quickly on one side would throw off sync on the other side. With this fix, even if some characters are skipped Flow will be able to later recover the matching of open and close tag.

Reviewed By: SamChou19815

Differential Revision: D51410168

fbshipit-source-id: 383b8b87a8de1fd3f28e499eb1a1d782ce4c4e95