[flow] Fix missing autocomplete results within components
Summary: This diff fixes the issue with autocomplete results within components. The issue was that the position of the scope containing the component's body doesn't overlap with the position of the component's body itself. I think this is because we didn't realize there was semantic uses of the scope location, but whatever the reason, this meant that autocomplete wouldn't suggest local variables within components when prompted.

Reviewed By: SamChou19815

Differential Revision: D51540985

fbshipit-source-id: 2f0f2f1f693f1cc46d8bf29748b20442bc849ca9