[flow] Always exclude libdefs from batch coverage
Summary:
Currently, we don't show coverage for libdef files as an accident, since they are not in the checked file and all files set. This will change in the future when we moved libdef validation into the checking phase. At that time, we could also choose to show libdef coverage.

However, since nobody seems to complain that we don't show libdef coverage, and I want to reduce difference in behavior between the current mode and `libdef_in_checking=true` mode, I decide to always exclude libdefs from coverage for now.

Changelog: [internal]

Reviewed By: gkz

Differential Revision: D51487150

fbshipit-source-id: 31f9f877db5fb17ee5feccf816c22c5bf40a71fa