[flow] Include libdef files in lazy stats
Summary:
In the future, when libdefs are always validated in the checking phase, it will naturally lead to the result of them being included in the checked set and all files set. I don't want to fork the .exp files just for these stats, so I decide to frontload the change.

In this diff, I always add the number of libdef files to both the numerator and denominator of the checked/total files fraction. This isn't lying in the current system. The libdef files are indeed always checked, just not during checking, but during InitLib phase.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D51487152

fbshipit-source-id: e9f8250c50489ec90d2b8d2e1074a739362a5ec6