[flow] Exclude libdef files from saved state files
Summary:
Libdef files are tricky to include in the saved state, because we unpack them in a temporarily directory, which might vary by machine. Currently, we are able to get away with this since libdefs are not in the files set in the server environment, but this will soon change once we move libdef validation into the checking phase.

In this diff, I did the preparation of always excluding them from saved state.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision:
D51500641

------------------------------------------------------------------------
(from 538d0634a98a2f50e6630acc9b5413c2e81c4b22)

fbshipit-source-id: 3c83502743abb3d639c32eb5711dc7f36a017529