[flow][conditional-type] Fix placeholder leaking issue in nested conditional type
Summary:
When we have unresolved inputs into conditional type during implicit instantiation, we will just produce a placeholder type (a variant of any) to signal that we don't know what the result is, so we will just put up the most permissive type possible, but doesn't make it contribute to solutions of implicit instantiation.

Placeholders are tricky to get right. In particular, we need to ensure that it doesn't leak. In hint decomposition, we will fail it if the final result contains placeholder. However, here in conditional type, we are not very careful so the effect leaked.

In the added test, we have two nested conditional type. During solve_targs of implicit instantiation, we will receive unresolved tvar from `T` and `K`. When we are evaluating `PropType`, we correctly produced a placeholder as a result. But then, the placeholder any propagated to the evaluation of `KeyType`, causing the true branch to be taken, but we should still refuse to evaluate it and return placeholder, so that we remain permissive and don't error, and the actual evaluation is deferred to instantiation with solution stage. However, since we picked the true branch, we will error, and speculation will fail on both branches, which leads to the confusing error in the GitHub issue.

The fix is fairly simple. We will defend against both unresolved inputs and placeholder inputs. In both cases, we don't know which branch to take, so we should produce placeholder. Placeholder in, placeholder out.

Close #9107.

Changelog: [fix] We fixed an issue that will cause some nested conditional type to be evaluated to any. (Example: Issue #9107)

Reviewed By: panagosg7

Differential Revision: D51644844

fbshipit-source-id: 3f9d5c274d933030493888532a5e24c30b6d882e