[flow] Conditionally take libdef into account for files API
Summary:
One major challenge of moving libdef validation into checking phase is that currently the parsing phase explicitly ignores libdef files in the `files.ml` API, so all the libdef files are never in the various filename set in server environment, so they never reach the checking phase.

This diff is a major preparation to remove that obstacle. I have changed the APIs in files.ml to optionally consider libdefs if named label parameters like `consider_libdef` or `include_libdef` is true. `include_libdef` includes libdef files, and `consider_libdef` will consider the `libs` set to correctly decide whether a js file is a normal file or libdef file.

In this diff, all these options are either explicitly set to false, or passed `Options.libdef_in_checking` which is an experimental flag that's always false for now.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D51487151

fbshipit-source-id: de47d7d00eecd61629b17c7830b22ea16b074a2b