[flow] Make `experimental.libdef_in_checking=true` actually move libdef validation into the checking phase
Summary:
After all previous preparation, this is the diff that finally does it!

When the flag is on, we first disable the validation in the InitLib phase. Then we start to include libdefs files in the mk_next_files for parsing, so that they will be parsed, added to dependency graph, and checking phase will see it. We have already let the checking phase automatically decide whether to check a file as a normal file or libdef file based on `File_key` variant, so that part will automatically work.

The above workflow works for a full check, but we need additional work to ensure lazy mode maintains the old behavior. Under lazy mode, after we finished init, we will run another check just for libdefs, so that we maintain the old invariant that libdefs are always validated during init.

Saved state is also a little tricky. D51500641 excluded libdef files from saved state, so we need to do additional work to restore the state for libdefs, by explicitly doing resolve_require and add them back to dependency graph. Fortunately, this is quite simple, since nothing depends on libdef in the current system. (All files do depend on libdefs, but they depend on the master_cx generated in InitLib, not the ones checked in checking phase.) Under the new system, InitLib will require libdef files to be parsed first. Therefore, we will also run an additional phase phase during saved state init before InitLib.

In this diff, I excluded LSP from the change, so that we still don't serve LSP requests from libdef files. However, with the infra change done in this diff, it will be relatively easy to add LSP support, so that will happen soon.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D51487149

fbshipit-source-id: d371e8daa78d70c76f913bfaa8d879fb5b9615db