[flow] Make `experimental.libdef_in_checking=true` enable LSP service for non-builtin libdef files
Summary:
In this diff, we enable LSP services when `experimental.libdef_in_checking=true` is true. The change is very straightforward: we just need to no longer ignore libdef files.

In principle, it can also work for builtin libdef files, but LSP clients are not even sending the requests to the server, since they are unpacked to temporary directory that are not watched.

Changelog: [internal]

Reviewed By: gkz

Differential Revision: D51503464

fbshipit-source-id: 7e6dbea0fcebb0771caf4cbb17752ccdf0cc788e