[flow] Extend DeepReadOnly data structures to allow for hook returns, other kinds of deep read onlyness
Summary:
In future diffs we'll be extending deep readonlyness to other values beyond component props. This diff prepares for this, and especially for making the result of calling a hook deep-read-only, by introducing a `dro_type` kind and threading it through the DRO constructors along with the location of the operation that triggers deep-read-onlyness. With this we can create specific error messages for the different kinds of deep-read-onlyness.

This diff also resturctures our test cases to move deep read only tests into the `react_rules` test case

Reviewed By: panagosg7

Differential Revision: D51435476

fbshipit-source-id: 146035691acb2d41474ae20ee82de05c6cffaaf3