[flow] Thread concrete AST node through annotations in name_def
Summary:
This diff modifies the `Annotation` name def to optionally contain the name def that it would have were it not for an annotation. This allows us to "look through" the annotation to see the expression that would initialize a name, if available.

For example, `const x: number = 42` produces an `Annotation` name def for `x`, but with this diff that name def also contains the `Value` name def containing the expression `42`.

This value name def is never (and should never be) typechecked. It exists to allow analyses to propagate information from initializers to variables even if an annotation is interposed.

This diff also exposes the name def structure as part of the loc env, so that analyses can consume it.

Reviewed By: panagosg7

Differential Revision: D51414943

fbshipit-source-id: 2c7bdbdd80636a158257cc0273235624ec6c57c4