[flow] Union of exact objects allow computed access as long as one of the members of the union have the property
Summary:
This already worked for named properties since D34408092, here we make it work for computed properties as well. This will be more common once we support numeric keys, as they always need to use computed access.

Changelog: [errors] Union of exact objects allow computed access as long as one of the members of the union have the property

Reviewed By: SamChou19815

Differential Revision: D51578298

fbshipit-source-id: a134633e4e6ad861cf866ad9d652127b5f1885b3