[flow] New nullish property refinement
Summary:
This diff creates a new env builder refinement for values that have a property determined to be nullish. This is very similar to the existing sentinel and `PropExists` refinements, but it is introduced by nonstrict `== null/void` tests.

Interestingly, there already exists a type-level implementation for this refinement, in the `PropNonMaybeP` predicate type. However, there is presently no way to introduce such predicates. This diff introduces them in the env builder.

While this does have an observable effect on typechecking, the motivation for this is actually to support react rule enforcement in D51233620.

Reviewed By: panagosg7

Differential Revision: D51233158

fbshipit-source-id: 657dfaa9e15f24960591c04400ccee5574f84c2a