[flow] Restore old behavior of libdef inclusion overriding ignores under `experimental.libdef_in_checking=true`
Summary:
Currently under the default config, we have the behavior that libdef inclusion will ignore whatever is specified in the ignore section. e.g. if we explicitly include `lib.js` in the `[libs]` section, then we won't honor `<PROJECT_ROOT>/lib.js` in the `[ignore]` section.

I don't know whether it's a intentional decision, but it's a result of the current system completely separating `SourceFile` and `LibFile`, and the ignore only applying to source files. As we move libdef validation into checking phase under `experimental.libdef_in_checking=true`, this is no longer true. I tried to align the behavior in most other cases, but missed this one.

As a result, a libdef file that is included and then ignored won't be parsed, but we still expect it to exist in during init_lib. init_lib will check whether a libfile is in parsing heap to decide whether it has parsing error. However, since we are in this inconsistent state, init_lib doesn't find it, so `lib_ok=false`, but no errors are printed to indicated something went wrong, since there is no parsing errors. When `lib_ok=false`, we just skip checking everything. Therefore, when we have a any lib file that is included in libs section but then ignored, we will just silently not checking anything! The added `libdef_conflicting_ignore` will demonstrate this bug before the fix in this diff.

The fix is fairly simple. Under `experimental.libdef_in_checking=true`, a file is wanted if it's not ignored OR it's a libdef file. Not AND.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D51728352

fbshipit-source-id: 747dc9198d8e884418c0f7c4c7e678f85127ee99